requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
alembic==1.14.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
URLObject==2.4.3
Werkzeug==3.1.3
pytesseract==0.3.10
Pillow==10.1.0